Put data point info in right-hand sidebar
