+++ src/components/plot-layouts/ScatterPlot/ScatterPlot.vue
@@ -93,7 +93,6 @@
                 </g>
             </svg>
         </div>
-        <DataPointInfo />
     </div>
 
 </template>
@@ -110,7 +109,6 @@ import CategoricFilter from "@/models/filters/CategoricFilter"
 
 // Components
 import ScatterPlotPointLayer from "./ScatterPlotPointLayer.vue"
-import DataPointInfo from "../../tools/overlays/DataPointInfo.vue"
 
 const dataStore = useDataStore()
 const scatterStore = useScatterStore()
