+++ src/components/plot-layouts/ScatterPlot/ScatterSideMenuRight.vue
@@ -1,0 +1,13 @@
+<template>
+    <div class="container side-menu-container">
+        <DataPointInfo />
+    </div>
+</template>
+
+<script setup>
+import DataPointInfo from '@/components/tools/DataPointInfo'
+</script>
+
+<style lang="scss" scoped>
+
+</style>
